APP-3921: Able to resubmit identity docs while there is already In-progress
