+++ buildSrc/src/main/kotlin/ApplicationConfiguration.kt
@@ -1,7 +1,7 @@
 object ApplicationConfiguration {
     const val artifactId = "java-lib-wallet-sdk"
     const val groupId = "com.paysera.lib"
-    const val version = "6.8.20"
+    const val version = "6.8.21-SNAPSHOT"
     const val name = "Java Library Wallet SDK"
     const val url = "https://www.paysera.com/v2/en-GB/index"
     const val description = "Paysera: Java Wallet SDK Library"
